fix: average rank

+++ app/src/main/java/com/goquestly/data/remote/dto/SessionStatisticsDto.kt
@@ -7,7 +7,7 @@ data class SessionStatisticsDto(
     val totalSessions: Int,
     val finishedSessions: Int,
     val finishRate: Double,
-    val averageRank: Int?,
+    val averageRank: Double?,
     val bestRank: Int?,
     val totalScore: Int,
     val totalCheckpointsPassed: Int,
